requirements.txt
numpy
libigl
pytest